home-server/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
openai-whisper>=20231117
eventlet>=0.33.0
